+++ src/ServiceProvider.php
@@ -5,9 +5,7 @@ namespace Butler\Auth;
 use Butler\Auth\Commands\GenerateSecretKey;
 use Butler\Auth\Commands\GenerateToken;
 use Illuminate\Contracts\Container\Container;
-use Illuminate\Foundation\Application as LaravelApplication;
 use Illuminate\Support\ServiceProvider as BaseServiceProvider;
-use Laravel\Lumen\Application as LumenApplication;
 
 class ServiceProvider extends BaseServiceProvider
 {
@@ -49,12 +47,10 @@ class ServiceProvider extends BaseServiceProvider
 
     protected function setupConfig(Container $app)
     {
-        $source = realpath($raw = __DIR__.'/../config/butler.php') ?: $raw;
+        $source = realpath($raw = __DIR__ . '/../config/butler.php') ?: $raw;
 
-        if ($app instanceof LaravelApplication && $app->runningInConsole()) {
+        if ($app->runningInConsole()) {
             $this->publishes([$source => config_path('butler.php')]);
-        } elseif ($app instanceof LumenApplication) {
-            $app->configure('butler');
         }
 
         $this->mergeConfigFrom($source, 'butler');